+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### ✅ Added
 - Image gallery
+- Editing messages
+- Mentions
+- Composer commands
+- Configuration of channel item swipe area
+
+### 🔄 Changed
+- Creation of channel items
 
 # [4.6.3](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.6.3)
 _December 21, 2021_